fixed betty issues, and pedantic errors

+++ 0x1A-hash_tables/5-hash_table_print.c
@@ -8,7 +8,7 @@
 void hash_table_print(const hash_table_t *ht)
 {
 	hash_node_t *tmp = NULL;
-	int i = 0;
+	unsigned long int i = 0;
 
 	if (ht == NULL)
 		return;
