+++ 0x1A-hash_tables/6-hash_table_delete.c
@@ -9,7 +9,7 @@ void hash_table_delete(hash_table_t *ht)
 {
 	hash_node_t *tmp = NULL;
 	hash_node_t *prev = NULL;
-	int i = 0;
+	unsigned long int i = 0;
 
 	if (ht == NULL)
 		return;
